Transport protocol v1 -- master
This adds CRC-16 checksums for the call_application() function. The
command from the master includes and additional packet containing the
CRC which the slave can verify. The status from the slave after an app
has processed the command included the CRC of the reply data and the
status message itself.

The protocol is backward compatible by detecting the v0 protocol
from the status response and ignoring any CRCs.

In the event of a CRC failure when reading the status or response data,
the read operation is retried 3 times. In the event of a CRC failure of
the request, the transaction is restarted a maximum of 3 times.

Bug: 66104849
Bug: 80133226
Test: TransportTest
Test: Update Android and unlock, update Citadel and can still unlock,
      reboot and unlock again
Change-Id: If0d37ee492ecddcc17c4dbde3ca992698a48a5e7