The Linux citadel_updater accepts a default device
The Linux version of citadel_updater accepts a --device arg to
specify the serial number of the FTDI device to use to talk to
Citadel.

This CL allows the environment variable CITADLE_DEVICE to hold
the default value of the --device arg. Of course, you can still
override that with the arg itself.

Change-Id: Ica963cd31ad746f93f6422ea43ccca7548ce063b
Signed-off-by: Bill Richardson <wfrichar@google.com>