Transport: improve logs for debugging
Add the app ID and details about the call so we can infer what the AP is
doing with Citadel. The app ID acts as a transaction ID as there will
only be one transaction with an app at a time.

This also fixes the logging priorities. Debug is a higher priority than
verbose and verbose is off by default in the Android build. The tracing
logs should be in debug and the extra spammy details are in verbose so
they can be enabled when needed.

Test: run on a phone and view the logs
Change-Id: I0a11e04c56ec64f8018c041055fe247deb633d9d