Transport: Add retryng mechanism while calling nos application.
Citadel chip sometimes give the response "APP_ERROR_TOO_MUCH", which
is to complain AP sent it a count different from what AP claimed or
more than it can accept - but this should not happen.

Give to the chip a little bit of time and retry calling again.

Also increase CRC_RETRY_COUNT from 3 to 5.

Bug: 147200739
Change-Id: I10308cecab382b29c7846d9b79ff3edf36534c8d
Signed-off-by: dybertwang <dybertwang@google.com>