proto: Add product certs ID for dauntless
Start from mini-proto build of dauntless, gChip phase in production
batch certificates which anchors to Android Root Certificate.
There are 2 set of production batch certificate for these pre-product
chips. The first set is used on mini-proto and proto-1.0 build, and the
second is used on proto-1.1 build.

Refer to go/pixel-dauntless-sign-strategy-so-far for more detail

Bug: 165285749
Change-Id: Ib3eca4b32eaec950389d5e8b4949882c1fa13010
Reviewed-on: https://nugget-os-review.googlesource.com/c/host/generic/+/38224
Reviewed-by: Bill Richardson <wfrichar@google.com>
Reviewed-by: Tommy Chiu <tommychiu@google.com>
Tested-by: Tommy Chiu <tommychiu@google.com>
Presubmit-Verified: TreeHugger Robot