Add NuggetClientDebuggable class for testing
This class allows tests that talk over Ultradebug to a dev board
to inspect the Request bytes sent to CallApp() and the Response
bytes returned from it without impacting the behavior. Test
frameworks can use this capability to check for data leaks due to
uninitialized data or stack overflow.

Bug: 175117199
Test: manual
Signed-off-by: Bill Richardson <wfrichar@google.com>
Change-Id: Ifa660c29e2d44b7ecd201926dfd9ea877692321a
Reviewed-on: https://nugget-os-review.googlesource.com/c/host/generic/+/38625
Reviewed-by: Tommy Chiu <tommychiu@google.com>
Presubmit-Verified: TreeHugger Robot