Fix struct nugget_app_storage for C++
The C99 way to have zero-length arrays in a struct is "flexible
array members", aka "variable length arrays". That's okay for C,
but it's not part of C++ (although Clang claims to allow it).

There are a few C++ programs that #include app_nugget.h, so we
have to fix it up a bit.

Bug: 173056823
Test: manual

Change-Id: I0fb3d4afb27b71b7b0f8050d4b1f51f4e118514f
Signed-off-by: Bill Richardson <wfrichar@google.com>
Reviewed-on: https://nugget-os-review.googlesource.com/c/host/generic/+/43864
Reviewed-by: Randall Spangler <rspangler@google.com>
Reviewed-by: Tommy Chiu <tommychiu@google.com>
Presubmit-Verified: TreeHugger Robot <android-build-prod@system.gserviceaccount.com>